fix: エラーメッセージ内部のインデントを2に修正し、ついでにinert_attrのパースについてparse.rsへの分割をして構造をわかりやすくした
